+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
-- Added CODEOWNERS file.
+- Added CODEOWNERS file;
+- Added resources for managing AWS Security Groups.
 
 ### Fixed